refactor - alteração no tipo de uma prop

+++ DTOs/Stock/FMPStock.cs
@@ -12,7 +12,7 @@ namespace api.DTOs
         public double beta { get; set; }
         public int volAvg { get; set; }
         public long mktCap { get; set; }
-        public int lastDiv { get; set; }
+        public double lastDiv { get; set; }
         public string range { get; set; } = string.Empty;
         public double changes { get; set; }
         public string companyName { get; set; } = string.Empty;
